feat(client): de optimizo el slider del dashboard
